+++ tests/unit/cli_test.py
@@ -7,11 +7,17 @@ import logging
 
 import unittest
 from wsgid.core.cli import Cli
-from wsgid.core import parser
+from wsgid.core import parser, WsgidApp
 from wsgid.commands import CommandInit
 from wsgid.test import fullpath, FakeOptions
+import wsgid.conf
 
-from mock import patch
+import daemon
+
+from mock import patch, Mock, MagicMock
+
+ROOT_PATH = fullpath(__file__)
+FIXTURES_PATH = os.path.join(ROOT_PATH, 'fixtures')
 
 class CliTest(unittest.TestCase):
 
@@ -22,13 +28,16 @@ class CliTest(unittest.TestCase):
     # so the tested applications does not try to use them
     sys.argv[1:] = []
     self.fake_app_path = os.path.join(fullpath(__file__), 'app-path')
-    
+
     # Ok, not pretty but better than re-implementing this in python
     os.system("rm -rf {0}".format(os.path.join(self.fake_app_path, 'pid/')))
     self.cli.options = parser._parse_args()
     self.cli.options.app_path = self.fake_app_path
 
     CommandInit().run(FakeOptions(app_path=self.fake_app_path))
+
+  def tearDown(self):
+      wsgid.conf.settings = None
 
   def test_nodaemon(self):
     opts = self._parse()
@@ -130,6 +139,7 @@ class CliTest(unittest.TestCase):
     self.assertEquals(True, options.debug)
     self.assertEquals(True, options.keep_alive) #If one option does now exist in the config file, we get the default value
     self.assertEquals(False, options.chroot)
+    self.assertEquals(True, options.stdout)
 
     self.assertEquals({'ENV1': 'VALUE1', 'ENV2': 'VALUE2'}, options.envs)
 
@@ -155,7 +165,7 @@ class CliTest(unittest.TestCase):
 
     self.cli._write_pid(42, self.cli.MASTER)
     self.cli._write_pid(43, self.cli.WORKER)
-    
+
     # Check we created all necessary paths
     self.assertTrue(os.path.exists(pid_folder))
     self.assertTrue(os.path.exists(master_pid_folder))
@@ -174,5 +184,70 @@ class CliTest(unittest.TestCase):
 
     self.cli._remove_pid(42, self.cli.MASTER)
     self.assertFalse(os.path.exists(pid_file))
+
+  '''
+   Doing this wsgid is able to run with a process supervisor.
+   See: issue/20 at github
+  '''
+  @patch('sys.stderr')
+  def test_start_workers_when_nodaemon(self, *args):
+    path = os.path.join(FIXTURES_PATH, 'nodaemon-app')
+    initcmd = CommandInit()
+    initcmd.run(FakeOptions(app_path=path))
+    with patch('daemon.DaemonContext', new=MagicMock()):
+      with patch.object(Cli, '_create_worker'):
+        with patch.object(Cli, 'validate_input_params'):
+          with patch.object(Cli, '_wait_workers'):
+            daemon.DaemonContext.__enter__ = Mock()
+            daemon.DaemonContext.__exit__ = Mock()
+            sys.argv[1:] = ['--workers=3', '--no-daemon', '--app-path={0}'.format(path)]
+            cli = Cli()
+            cli.run()
+            self.assertEquals(3, cli._create_worker.call_count)
+            self.assertEquals(1, cli._wait_workers.call_count)
+
+  '''
+   When using only --do-daemon (without --stdout) logs must be
+   redirected to log file.
+  '''
+  def test_daemon_should_create_normal_logs(self):
+    opts = FakeOptions(app_path="/some/path", no_daemon=True, stdout=False, debug=None, chroot=None)
+    with patch('logging.FileHandler'):
+      cli = Cli()
+      cli._set_loggers(opts)
+      self.assertEquals(1, logging.FileHandler.call_count)
+      self.assertEquals((("/some/path/logs/wsgid.log",), {}), logging.FileHandler.call_args)
+
+  '''
+   When we pass --stdout all logs must be pointed to stdout.
+  '''
+  def test_create_log_for_stdout(self):
+    opts = FakeOptions(app_path="/some/path", no_daemon=True, stdout=True, debug=None, chroot=None)
+    with patch('logging.StreamHandler'):
+      cli = Cli()
+      cli._set_loggers(opts)
+      self.assertEquals(1, logging.StreamHandler.call_count)
+
+  def test_clean_pid_files_on_keyboard_interrupt(self):
+    path = os.path.join(FIXTURES_PATH, 'clean-pids-app')
+    initcmd = CommandInit()
+    opts = FakeOptions(app_path=path)
+    initcmd.run(opts)
+    open(os.path.join(path, 'pid/master/3340.pid'), 'w')
+    open(os.path.join(path, 'pid/worker/2736.pid'), 'w')
+    open(os.path.join(path, 'pid/worker/3847.pid'), 'w')
+    with patch('os.wait'):
+      with patch('os.getpid'):
+          os.getpid.side_effect = lambda: 3340
+          os.wait.side_effect = KeyboardInterrupt()
+          cli = Cli()
+          cli.options = opts
+          cli.log = Mock()
+          cli.workers = [2736, 3847]
+          cli._wait_workers()
+          app = WsgidApp(path)
+          self.assertEquals([], app.master_pids())
+          self.assertEquals([], app.worker_pids())
+