+++ wsgid/commands/__init__.py
@@ -3,3 +3,4 @@
 from init import *
 from config import *
 from manage import *
+from status import *