+++ wsgid/commands/config.py
@@ -29,12 +29,15 @@ class CommandConfig(Plugin):
       self._override_if_not_none('workers', cfg_values, options.workers, convert_func=int)
     self._override_if_not_none('keep_alive', cfg_values, options.keep_alive)
     self._override_if_not_none('chroot', cfg_values, options.chroot)
+    self._override_if_not_none('no_daemon', cfg_values, options.no_daemon)
     self._override_if_not_none('recv', cfg_values, options.recv)
     self._override_if_not_none('send', cfg_values, options.send)
-    
+    self._override_if_not_none('mongrel2_chroot', cfg_values, options.mongrel2_chroot)
+
     # Custom config command options
-    cfg_values['debug'] = str((not options.no_debug))
-    
+    if options.no_debug:
+        cfg_values['debug'] = str((not options.no_debug))
+
     # Rewrite the config file
     f.seek(0)
     f.truncate()
@@ -42,7 +45,7 @@ class CommandConfig(Plugin):
     f.close()
 
   def extra_options(self):
-    return [CommandLineOption(name='no-debug', help = 'Turns off debug option', type=BOOL),
+    return [CommandLineOption(name='no-debug', help = 'Turns off debug option', type=BOOL, default_value=None),
            CommandLineOption(name='no-keep-alive', help = 'Turns off Keep alive option', type=BOOL),
            CommandLineOption(name='no-chroot', help = 'Turns off Chroot option', type=BOOL)]