+++ wsgid/core/wsgid.py
@@ -6,14 +6,16 @@ from . import StartResponse, get_main_logger
 import zmq
 from StringIO import StringIO
 import sys
+import os
+from .. import conf
 
 class Wsgid(object):
-  
+
   def __init__(self, app = None, recv = None, send = None):
     self.app = app
     self.recv = recv
     self.send = send
-    
+
     self.ctx = zmq.Context()
     self.log = get_main_logger()
 
@@ -37,12 +39,24 @@ class Wsgid(object):
         self.log.debug("Disconnect message received, id=%s" % m2message.client_id)
         continue
 
+      if m2message.is_upload_start():
+        self.log.debug("Starting async upload, file will be at: {0}".format(m2message.async_upload_path))
+        continue
+
 
       # Call the app and send the response back to mongrel2
       self._call_wsgi_app(m2message, send_sock)
-      
+
   def _call_wsgi_app(self, m2message, send_sock):
     environ = self._create_wsgi_environ(m2message.headers, m2message.body)
+    upload_path = conf.settings.mongrel2_chroot or '/'
+
+    if m2message.is_upload_done():
+        self.log.debug("Async upload done, reading from {0}".format(m2message.async_upload_path))
+        parts = m2message.async_upload_path.split('/')
+        upload_path = os.path.join(upload_path, *parts)
+        environ['wsgi.input'] = open(upload_path)
+
     start_response = StartResponse()
 
     server_id = m2message.server_id
@@ -68,6 +82,15 @@ class Wsgid(object):
     finally:
       if hasattr(response, 'close'):
         response.close()
+      if m2message.is_upload_done():
+        self._remove_tmp_file(upload_path)
+
+  def _remove_tmp_file(self, filepath):
+      try:
+        os.unlink(filepath)
+      except OSError, o:
+        self.log.exception(o)
+
 
   '''
    Constructs a mongrel2 response message based on the
@@ -101,7 +124,7 @@ class Wsgid(object):
     environ = {}
     #Not needed
     json_headers.pop('URI', None)
-    
+
     #First, some fixed values
     environ['wsgi.multithread'] = False
     environ['wsgi.multiprocess'] = True
@@ -109,6 +132,7 @@ class Wsgid(object):
     environ['wsgi.errors'] = sys.stderr
     environ['wsgi.version'] = (1,0)
     self._set(environ, 'wsgi.url_scheme', "http")
+
 
     if body:
       environ['wsgi.input'] = StringIO(body)
@@ -137,10 +161,10 @@ class Wsgid(object):
     self._set(environ, 'SERVER_NAME', server_name)
 
     self._set(environ, 'REMOTE_ADDR', json_headers['x-forwarded-for'])
-    
+
     self._set(environ, 'CONTENT_TYPE', json_headers.pop('content-type', ''))
     environ['content-type'] = environ['CONTENT_TYPE']
-    
+
     self._set(environ, 'CONTENT_LENGTH', json_headers.pop('content-length', ''))
     environ['content-length'] = environ['CONTENT_LENGTH']
 
@@ -160,4 +184,5 @@ class Wsgid(object):
   '''
   def _set(self, environ, key, value):
     environ[key] = str(value)
+